requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==2.3.7
RPi.GPIO==0.7.1
python-dotenv==1.0.0
